refactor: separate theme toggle logic & styles
